# docs/swagger.yaml
definitions:
  common.ExceptionModel:
    properties:
      code:
        type: integer
      message:
        type: string
      stackTrace:
        type: string
    type: object
  common.ResponseModel-dto_UserDto-dto_UserDto:
    properties:
      exceptionModel:
        $ref: '#/definitions/common.ExceptionModel'
      input:
        $ref: '#/definitions/dto.UserDto'
      output:
        $ref: '#/definitions/dto.UserDto'
      responseCode:
        type: integer
    type: object
  dto.UserDto:
    properties:
      age:
        type: integer
      lastname:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/redis/{key}:
    get:
      consumes:
      - application/json
      description: get value by key
      parameters:
      - description: key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Show value for key
  /api/v1/redis/{key}/{value}:
    put:
      consumes:
      - application/json
      description: put key and value
      parameters:
      - description: key
        in: path
        name: key
        required: true
        type: string
      - description: value
        in: path
        name: value
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: put key and value
  /api/v1/user/:
    post:
      consumes:
      - application/json
      description: save user
      parameters:
      - description: User Dto
        in: body
        name: UserDto
        required: true
        schema:
          $ref: '#/definitions/dto.UserDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.ResponseModel-dto_UserDto-dto_UserDto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ResponseModel-dto_UserDto-dto_UserDto'
      summary: save user
  /api/v1/user/{id}:
    get:
      consumes:
      - application/json
      description: find user by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserDto'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: find user by id
swagger: "2.0"